refactor: move types files into types folder
