+++ packages/ai-proxy/src/types/errors.ts
