+++ packages/ai-proxy/src/types/mcp-config-checker.ts
@@ -1,6 +1,6 @@
-import type { McpConfiguration } from './mcp-client';
+import type { McpConfiguration } from '../mcp-client';
 
-import McpClient from './mcp-client';
+import McpClient from '../mcp-client';
 
 export default class McpConfigChecker {
   static check(mcpConfig: McpConfiguration) {
